heredoc + helpers norm compliant
